+++ FranAlonso/Features/Clients/Domain/UseCases/ObserveClientsUseCase.swift
@@ -1,11 +1,13 @@
 struct ObserveClientsUseCase: Sendable {
-    private let repository: any ClientRepository
-
-    init(repository: any ClientRepository) {
-        self.repository = repository
-    }
+    private let clientRepository: any ClientRepository
 
     func callAsFunction() async -> AsyncThrowingStream<[Client], any Error> {
-        await repository.observeClients()
+        await clientRepository.observeClients()
+    }
+}
+
+extension ObserveClientsUseCase {
+    init(repository: any ClientRepository) {
+        self.init(clientRepository: repository)
     }
 }