+++ FranAlonso/Features/Clients/Presentation/Screens/ClientListScreen.swift
@@ -4,17 +4,19 @@ import SwiftUI
 struct ClientListScreen: View {
     @State private var viewModel: ClientListViewModel
 
-    init(observeClients: ObserveClientsUseCase) {
-        _viewModel = State(
-            initialValue: ClientListViewModel(observeClients: observeClients)
-        )
-    }
-
     var body: some View {
         ClientListContent(state: viewModel.state)
             .navigationTitle(Text(.clientsListTitle))
             .task {
                 await viewModel.load()
             }
+    }
+}
+
+extension ClientListScreen {
+    init(observeClients: ObserveClientsUseCase) {
+        _viewModel = State(
+            initialValue: ClientListViewModel(observeClients: observeClients)
+        )
     }
 }